Refactor: move JsonParser to separate file (#4586)
